Add files via upload
The animation starts with pipes being layered from the drilling rig into the shale rock layer. Once the pipe reaches the shale rock, the chemicals allow trapped gas and oil in the shale rock layer to flow into the pipes. Fractures are formed on the sides of the pipes, and the chemicals from the pipes start to leach into the soil shown by liquid flow. Next, the chemicals start to travel up into the groundwater, so the stonefly leaves before the chemicals mix with the groundwater and turn it orange. At the same time, the once trapped gases are accidentally released in the underground burrow of the larvae located on top of the shale layer. While the larvae tries to escape, the gases eventually suffocate the larvae. Eventually, the chemicals reach the plants and kills them, evidenced by a color change. As the plants die, you see a multitude of animals escape to find a new habitat as their old habitat can no longer support them. First, ants (which are now contaminated with chemicals), leave their ant burrow underground. Second, the butterfly flies upwards. Next, a fly resting on the rock follows the butterfly and flies up. Lastly, a blue jay, which was resting in its treehouse on the tree, travels left and leaves its home. Although the habitat is now destroyed and there are no more animals left, the gases continue to be imported into the pipe, which is representative of the long-term effects of fracking.